Remove v0.8 shell, and move 0.9 to its location.
