+++ samples/client/lit/package.json
@@ -13,15 +13,11 @@
     "serve:agent:contact_lookup": "cd ../../agent/adk/contact_lookup && uv run .",
     "serve:agent:orchestrator": "cd ../../agent/adk/orchestrator && uv run .",
     "serve:shell": "cd shell && npm run dev",
-    "serve:shell09": "cd shell_v0_9 && npm run dev",
     "serve:gallery09": "cd gallery_v0_9 && npm install && npm run dev",
     "build:renderer": "cd ../../../renderers && for dir in 'web_core' 'markdown/markdown-it' 'lit'; do (cd \"$dir\" && npm install && npm run build); done",
     "demo:restaurant": "npm install && npm run build:renderer && concurrently -k -n \"SHELL,REST\" -c \"magenta,blue\" \"npm run serve:shell\" \"npm run serve:agent:restaurant\"",
-    "demo:restaurant09": "npm install && npm run build:renderer && concurrently -k -n \"SHELL09,REST\" -c \"magenta,blue\" \"npm run serve:shell09\" \"npm run serve:agent:restaurant\"",
     "demo:contact": "npm install && npm run build:renderer && concurrently -k -n \"SHELL,CONT1\" -c \"magenta,green\" \"npm run serve:shell\" \"npm run serve:agent:contact_lookup\"",
-    "demo:contact09": "npm install && npm run build:renderer && concurrently -k -n \"SHELL09,CONT1\" -c \"magenta,green\" \"npm run serve:shell09\" \"npm run serve:agent:contact_lookup\"",
     "demo:orchestrator": "npm install && npm run build:renderer && concurrently -k -n \"SHELL,ORCH\" -c \"magenta,cyan\" \"npm run serve:shell\" \"npm run serve:agent:orchestrator\"",
-    "demo:orchestrator09": "npm install && npm run build:renderer && concurrently -k -n \"SHELL09,ORCH\" -c \"magenta,cyan\" \"npm run serve:shell09\" \"npm run serve:agent:orchestrator\"",
     "demo:gallery09": "npm install && npm run build:renderer && concurrently -k -n \"SHELL,GALLERY09\" -c \"magenta,yellow\" \"npm run serve:gallery09\""
   },
   "devDependencies": {
