+++ samples/client/lit/shell/README.md
@@ -4,47 +4,44 @@ This is a UI to generate and visualize A2UI responses.
 
 ## Prerequisites
 
-1. [nodejs](https://nodejs.org/en)
+* [nodejs](https://nodejs.org/en)
+* [uv](https://docs.astral.sh/uv/getting-started/installation/)
 
 ## Running
 
-This sample depends on the Lit renderer. Before running this sample, you need to build the renderer.
+### Configure Gemini API Key
 
-1. **Build the renderer:**
-   ```bash
-   cd ../../../renderers/web_core
-   npm install
-   npm run build
-   cd ../lit
-   npm install
-   npm run build
-   ```
+This step is **required** for the backend to work.
 
-2. **Run this sample:**
-   ```bash
-   cd - # back to the sample directory
-   npm install
-   ```
+- Navigate to the respective agent directory (e.g., `samples/agent/adk/restaurant_finder`
+  or `samples/agent/adk/contact_lookup`).
+- Copy `.env.example` to `.env` and set your `GEMINI_API_KEY`.
 
-3. **Run the servers:**
-   - Run the [Restaurant Finder Agent](../../../agent/adk/restaurant_finder/) (Default): `npm run demo:restaurant`
-   - Run the dev server: `npm run dev`
+(More details can be found in the READMEs of the [agent samples](/samples/agent/adk).)
 
-### Running the Contact Sample
+### Run the Demo and Servers
 
-The shell app supports multiple configured applications. To run the Contact sample:
+From the `samples/client/lit` directory:
 
-1. **Start the Contact Agent:**
-   ```bash
-   npm run demo:contact
-   ```
+- Run the **Restaurant Finder** demo (starts both the agent and the shell):
+  ```bash
+  npm run demo:restaurant09
+  ```
+- Or run the **Contact Lookup** demo:
+  ```bash
+  npm run demo:contact09
+  ```
 
-2. **Open the Contact App:**
-   - Open `http://localhost:5173/?app=contacts`
+### Open the Application
+
+From a web browser:
+
+- Open `http://localhost:5173/` for the default application (Restaurant Finder).
+- Open `http://localhost:5173/?app=contacts` for the Contact Lookup application.
 
 > **Note:** The `?app=` query parameter only supports apps that are actively configured in `app.ts` (e.g., `restaurant`, `contacts`). You cannot run arbitrary agents by passing their URL as a query string without first adding them to the shell configuration.
 
-After starting the dev server, you can open http://localhost:5173/ to view the sample.
+## Security Notice
 
 Important: The sample code provided is for demonstration purposes and illustrates the mechanics of A2UI and the Agent-to-Agent (A2A) protocol. When building production applications, it is critical to treat any agent operating outside of your direct control as a potentially untrusted entity.
