+++ samples/client/lit/shell/app.ts
@@ -21,64 +21,39 @@ import {
   html,
   css,
   nothing,
-  HTMLTemplateResult,
-  unsafeCSS,
 } from "lit";
 import { customElement, state } from "lit/decorators.js";
-import { theme as uiTheme } from "./theme/default-theme.js";
-import { A2UIClient } from "./client.js";
-import {
-  SnackbarAction,
-  SnackbarMessage,
-  SnackbarUUID,
-  SnackType,
-} from "./types/types.js";
-import { type Snackbar } from "./ui/snackbar.js";
 import { repeat } from "lit/directives/repeat.js";
-import * as v0_8 from "@a2ui/lit/v0_8";
-import * as UI from "@a2ui/lit/ui";
 
-// App elements.
-import "./ui/ui.js";
-
-// Configurations
-import { AppConfig } from "./configs/types.js";
-import { config as restaurantConfig } from "./configs/restaurant.js";
-import { config as contactsConfig } from "./configs/contacts.js";
-import { styleMap } from "lit/directives/style-map.js";
+// A2UI
+import * as v0_9 from "@a2ui/web_core/v0_9";
+import { basicCatalog, Context } from "@a2ui/lit/v0_9";
 import { renderMarkdown } from "@a2ui/markdown-it";
 
-const configs: Record<string, AppConfig> = {
-  restaurant: restaurantConfig,
-  contacts: contactsConfig,
-};
+// Configurations
+import { A2UIClient } from "./client.js";
+import { restaurantConfig, AppConfig } from "./configs/configs.js";
+import { styleMap } from "lit/directives/style-map.js";
 
 @customElement("a2ui-shell")
 export class A2UILayoutEditor extends SignalWatcher(LitElement) {
-  @provide({ context: UI.Context.theme })
-  accessor theme: v0_8.Types.Theme = uiTheme;
-
-  @provide({ context: UI.Context.markdown })
-  accessor markdownRenderer: v0_8.Types.MarkdownRenderer = renderMarkdown;
+  @provide({ context: Context.markdown })
+  accessor markdownRenderer: any = renderMarkdown;
 
   @state()
   accessor #requesting = false;
 
   @state()
-  accessor #error: string | null = null;
+  accessor #lastMessages: any[] = [];
 
   @state()
-  accessor #lastMessages: v0_8.Types.ServerToClientMessage[] = [];
-
-  @state()
-  accessor config: AppConfig = configs.contacts;
+  accessor config: AppConfig = restaurantConfig;
 
   @state()
   accessor #loadingTextIndex = 0;
   #loadingInterval: number | undefined;
 
   static styles = [
-    unsafeCSS(v0_8.Styles.structuralStyles),
     css`
       * {
         box-sizing: border-box;
@@ -160,6 +135,21 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
         }
       }
 
+      .material-symbols {
+        font-family: "Material Symbols Outlined", sans-serif;
+        font-variation-settings: "FILL" 1;
+        font-weight: normal;
+        font-style: normal;
+        font-size: 24px;
+        line-height: 1;
+        letter-spacing: normal;
+        text-transform: none;
+        display: inline-block;
+        white-space: nowrap;
+        word-wrap: normal;
+        direction: ltr;
+      }
+
       .rotate {
         animation: rotate 1s linear infinite;
       }
@@ -202,7 +192,8 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
         height: 48px;
         font-size: 32px;
 
-        & .g-icon {
+        & .material-symbols {
+          font-family: "Material Symbols Outlined";
           pointer-events: none;
 
           &::before {
@@ -212,7 +203,7 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
       }
 
       @container style(--color-scheme: dark) {
-        .theme-toggle .g-icon::before {
+        .theme-toggle .material-symbols::before {
           content: "light_mode";
           color: var(--n-90);
         }
@@ -270,38 +261,41 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
     `,
   ];
 
-  #processor = v0_8.Data.createSignalA2uiMessageProcessor();
+  // Create a Message Processor that uses the basic catalog.
+  #processor = new v0_9.MessageProcessor(
+    [basicCatalog],
+    async (action: v0_9.A2uiClientAction): Promise<any> => {
+      console.debug("Handling action", action);
+
+      const context: Record<string, any> = { ...action.context };
+
+      // Do we need to update this to a more strict v0.9 type?
+      const message = {
+        userAction: {
+          name: action.name,
+          surfaceId: action.surfaceId,
+          sourceComponentId: action.sourceComponentId,
+          timestamp: new Date().toISOString(),
+          context,
+        },
+      };
+
+      await this.#sendAndProcessMessage(message);
+    },
+  );
   #a2uiClient = new A2UIClient();
-  #snackbar: Snackbar | undefined = undefined;
-  #pendingSnackbarMessages: Array<{
-    message: SnackbarMessage;
-    replaceAll: boolean;
-  }> = [];
-
-  #maybeRenderError() {
-    if (!this.#error) return nothing;
-
-    return html`<div class="error">${this.#error}</div>`;
-  }
 
   connectedCallback() {
     super.connectedCallback();
 
-    // Load config from URL
-    const urlParams = new URLSearchParams(window.location.search);
-    const appKey = urlParams.get("app");
-    this.config = (appKey && configs[appKey]) || configs.contacts;
-
-    // Apply the theme directly, which will use the Lit context.
-    if (this.config.theme) {
-      this.theme = this.config.theme;
+    // Set the CSS Overrides for the given appKey.
+    if (this.config.cssOverrides && !document.adoptedStyleSheets.includes(this.config.cssOverrides)) {
+      document.adoptedStyleSheets = [
+        ...document.adoptedStyleSheets,
+        this.config.cssOverrides,
+      ];
     }
-
-    window.document.title = this.config.title;
-    window.document.documentElement.style.setProperty(
-      "--background",
-      this.config.background
-    );
+    document.title = this.config.title;
 
     // Initialize client with configured URL
     this.#a2uiClient = new A2UIClient(this.config.serverUrl);
@@ -312,7 +306,6 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
       this.#renderThemeToggle(),
       this.#maybeRenderForm(),
       this.#maybeRenderData(),
-      this.#maybeRenderError(),
     ];
   }
 
@@ -320,19 +313,19 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
     return html` <div>
       <button
         @click=${(evt: Event) => {
-        if (!(evt.target instanceof HTMLButtonElement)) return;
-        const { colorScheme } = window.getComputedStyle(evt.target);
-        if (colorScheme === "dark") {
-          document.body.classList.add("light");
-          document.body.classList.remove("dark");
-        } else {
-          document.body.classList.add("dark");
-          document.body.classList.remove("light");
-        }
-      }}
+          if (!(evt.target instanceof HTMLButtonElement)) return;
+          const { colorScheme } = window.getComputedStyle(evt.target);
+          if (colorScheme === "dark") {
+            document.body.classList.add("light");
+            document.body.classList.remove("dark");
+          } else {
+            document.body.classList.add("dark");
+            document.body.classList.remove("light");
+          }
+        }}
         class="theme-toggle"
       >
-        <span class="g-icon filled-heavy"></span>
+        <span class="material-symbols"></span>
       </button>
     </div>`;
   }
@@ -341,7 +334,7 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
     if (this.#requesting) return nothing;
     if (this.#lastMessages.length > 0) return nothing;
 
-    return html` <form
+    return html`<form
       @submit=${async (evt: Event) => {
         evt.preventDefault();
         if (!(evt.target instanceof HTMLFormElement)) {
@@ -352,17 +345,18 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
         if (!body) {
           return;
         }
-        const message = body as v0_8.Types.A2UIClientEventMessage;
+        const message = body as any;
         await this.#sendAndProcessMessage(message);
       }}
     >
       ${this.config.heroImage
         ? html`<div
             style=${styleMap({
-          "--background-image-light": `url(${this.config.heroImage})`,
-          "--background-image-dark": `url(${this.config.heroImageDark ?? this.config.heroImage
-            })`,
-        })}
+              "--background-image-light": `url(${this.config.heroImage})`,
+              "--background-image-dark": `url(${
+                this.config.heroImageDark ?? this.config.heroImage
+              })`,
+            })}
             id="hero-img"
           ></div>`
         : nothing}
@@ -378,7 +372,7 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
           ?disabled=${this.#requesting}
         />
         <button type="submit" ?disabled=${this.#requesting}>
-          <span class="g-icon filled-heavy">send</span>
+          <span class="material-symbols">send</span>
         </button>
       </div>
     </form>`;
@@ -386,14 +380,14 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
 
   #startLoadingAnimation() {
     if (
-      Array.isArray(this.config.loadingText) &&
+      this.config.loadingText &&
       this.config.loadingText.length > 1
     ) {
       this.#loadingTextIndex = 0;
       this.#loadingInterval = window.setInterval(() => {
         this.#loadingTextIndex =
           (this.#loadingTextIndex + 1) %
-          (this.config.loadingText as string[]).length;
+          this.config.loadingText!.length;
       }, 2000);
     }
   }
@@ -405,9 +399,7 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
     }
   }
 
-  async #sendMessage(
-    message: v0_8.Types.A2UIClientEventMessage
-  ): Promise<v0_8.Types.ServerToClientMessage[]> {
+  async #sendMessage(message: any): Promise<any[]> {
     try {
       this.#requesting = true;
       this.#startLoadingAnimation();
@@ -418,7 +410,7 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
 
       return response;
     } catch (err) {
-      this.snackbar(err as string, SnackType.ERROR);
+      console.error(err);
     } finally {
       this.#requesting = false;
       this.#stopLoadingAnimation();
@@ -429,14 +421,9 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
 
   #maybeRenderData() {
     if (this.#requesting) {
-      let text = "Awaiting an answer...";
-      if (this.config.loadingText) {
-        if (Array.isArray(this.config.loadingText)) {
-          text = this.config.loadingText[this.#loadingTextIndex];
-        } else {
-          text = this.config.loadingText;
-        }
-      }
+      const text = this.config.loadingText
+        ? this.config.loadingText[this.#loadingTextIndex]
+        : "Awaiting an answer...";
 
       return html` <div class="pending">
         <div class="spinner"></div>
@@ -444,121 +431,39 @@ export class A2UILayoutEditor extends SignalWatcher(LitElement) {
       </div>`;
     }
 
-    const surfaces = this.#processor.getSurfaces();
-    if (surfaces.size === 0) {
+    const surfaces = Array.from(this.#processor.model.surfacesMap.entries());
+    if (surfaces.length === 0) {
       return nothing;
     }
+    console.debug("Rendering surfaces", surfaces);
 
     return html`<section id="surfaces">
       ${repeat(
-      this.#processor.getSurfaces(),
-      ([surfaceId]) => surfaceId,
-      ([surfaceId, surface]) => {
-        return html`<a2ui-surface
-              @a2uiaction=${async (
-          evt: v0_8.Events.StateEvent<"a2ui.action">
-        ) => {
-            const [target] = evt.composedPath();
-            if (!(target instanceof HTMLElement)) {
-              return;
-            }
-
-            const context: v0_8.Types.A2UIClientEventMessage["userAction"]["context"] =
-              {};
-            if (evt.detail.action.context) {
-              const srcContext = evt.detail.action.context;
-              for (const item of srcContext) {
-                if (item.value.literalBoolean) {
-                  context[item.key] = item.value.literalBoolean;
-                } else if (item.value.literalNumber) {
-                  context[item.key] = item.value.literalNumber;
-                } else if (item.value.literalString) {
-                  context[item.key] = item.value.literalString;
-                } else if (item.value.path) {
-                  const path = this.#processor.resolvePath(
-                    item.value.path,
-                    evt.detail.dataContextPath
-                  );
-                  const value = this.#processor.getData(
-                    evt.detail.sourceComponent,
-                    path,
-                    surfaceId
-                  );
-                  context[item.key] = value;
-                }
-              }
-            }
-
-            const message: v0_8.Types.A2UIClientEventMessage = {
-              userAction: {
-                name: evt.detail.action.name,
-                surfaceId,
-                sourceComponentId: target.id,
-                timestamp: new Date().toISOString(),
-                context,
-              },
-            };
-
-            await this.#sendAndProcessMessage(message);
-          }}
-              .surfaceId=${surfaceId}
+        surfaces,
+        ([surfaceId]) => surfaceId,
+        ([_, surface]) => {
+          return html`<a2ui-surface
               .surface=${surface}
-              .processor=${this.#processor}
-            ></a2-uisurface>`;
-      }
-    )}
+            ></a2ui-surface>`;
+        },
+      )}
     </section>`;
   }
 
   async #sendAndProcessMessage(request) {
     const messages = await this.#sendMessage(request);
-
-    console.log(messages);
+    console.debug("Received messages", messages);
 
     this.#lastMessages = messages;
-    this.#processor.clearSurfaces();
+
+    // this.#processor.clearSurfaces();
+    // Why? Shouldn't `deleteSurface` be sent from the agent to the client?
+    for (const surfaceId of Array.from(
+      this.#processor.model.surfacesMap.keys(),
+    )) {
+      this.#processor.model.deleteSurface(surfaceId);
+    }
+
     this.#processor.processMessages(messages);
-  }
-
-  snackbar(
-    message: string | HTMLTemplateResult,
-    type: SnackType,
-    actions: SnackbarAction[] = [],
-    persistent = false,
-    id = globalThis.crypto.randomUUID(),
-    replaceAll = false
-  ) {
-    if (!this.#snackbar) {
-      this.#pendingSnackbarMessages.push({
-        message: {
-          id,
-          message,
-          type,
-          persistent,
-          actions,
-        },
-        replaceAll,
-      });
-      return;
-    }
-
-    return this.#snackbar.show(
-      {
-        id,
-        message,
-        type,
-        persistent,
-        actions,
-      },
-      replaceAll
-    );
-  }
-
-  unsnackbar(id?: SnackbarUUID) {
-    if (!this.#snackbar) {
-      return;
-    }
-
-    this.#snackbar.hide(id);
   }
 }
