+++ samples/client/lit/shell/client.ts
@@ -16,14 +16,13 @@
 
 import { Part, SendMessageSuccessResponse, Task } from "@a2a-js/sdk";
 import { A2AClient } from "@a2a-js/sdk/client";
-import { v0_8 } from "@a2ui/lit";
+import * as v0_9 from "@a2ui/web_core/v0_9";
 
 const A2UI_MIME_TYPE = "application/json+a2ui";
 
 export class A2UIClient {
   #serverUrl: string;
   #client: A2AClient | null = null;
-  #contextId?: string;
 
   constructor(serverUrl: string = "") {
     this.#serverUrl = serverUrl;
@@ -44,7 +43,10 @@ export class A2UIClient {
         {
           fetchImpl: async (url, init) => {
             const headers = new Headers(init?.headers);
-            headers.set("X-A2A-Extensions", "https://a2ui.org/a2a-extension/a2ui/v0.8");
+            headers.set(
+              "X-A2A-Extensions",
+              "https://a2ui.org/a2a-extension/a2ui/v0.9",
+            );
             return fetch(url, { ...init, headers });
           }
         }
@@ -54,10 +56,9 @@ export class A2UIClient {
   }
 
   async send(
-    message: v0_8.Types.A2UIClientEventMessage | string
-  ): Promise<v0_8.Types.ServerToClientMessage[]> {
+    message: any | string
+  ): Promise<any[]> {
     const client = await this.#getClient();
-
     let parts: Part[] = [];
 
     if (typeof message === 'string') {
@@ -87,7 +88,6 @@ export class A2UIClient {
     const response = await client.sendMessage({
       message: {
         messageId: crypto.randomUUID(),
-        contextId: this.#contextId,
         role: "user",
         parts: parts,
         kind: "message",
@@ -99,19 +99,11 @@ export class A2UIClient {
     }
 
     const result = (response as SendMessageSuccessResponse).result as Task;
-    
-    // Extract contextId
-    if (result.contextId) {
-      this.#contextId = result.contextId;
-    } else if ("contextId" in response && response.contextId) {
-      this.#contextId = response.contextId as string;
-    }
-
     if (result.kind === "task" && result.status.message?.parts) {
-      const messages: v0_8.Types.ServerToClientMessage[] = [];
+      const messages: any[] = [];
       for (const part of result.status.message.parts) {
         if (part.kind === 'data') {
-          messages.push(part.data as v0_8.Types.ServerToClientMessage);
+          messages.push(part.data);
         }
       }
       return messages;
