--- samples/client/lit/shell/theme/clone-default-theme.ts
@@ -1,22 +1,0 @@
-/*
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { theme } from "./default-theme.js";
-import { v0_8 } from "@a2ui/lit";
-
-export function cloneDefaultTheme(): v0_8.Types.Theme {
-  return structuredClone(theme);
-}
