--- samples/client/lit/shell/types/types.ts
@@ -1,42 +1,0 @@
-/*
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { HTMLTemplateResult } from "lit";
-
-export enum SnackType {
-  NONE = "none",
-  INFORMATION = "information",
-  WARNING = "warning",
-  ERROR = "error",
-  PENDING = "pending",
-}
-
-export type SnackbarUUID = ReturnType<typeof globalThis.crypto.randomUUID>;
-
-export type SnackbarAction = {
-  title: string;
-  action: string;
-  value?: HTMLTemplateResult | string;
-  callback?: () => void;
-};
-
-export type SnackbarMessage = {
-  id: SnackbarUUID;
-  type: SnackType;
-  persistent: boolean;
-  message: string | HTMLTemplateResult;
-  actions?: SnackbarAction[];
-};
