--- samples/client/lit/shell_v0_9/package.json
@@ -1,89 +1,0 @@
-{
-  "name": "@a2ui/shell_v0_9",
-  "private": true,
-  "version": "0.9.0",
-  "description": "A2UI Universal Shell (v0.9)",
-  "main": "./dist/shell.js",
-  "types": "./dist/shell.d.ts",
-  "type": "module",
-  "scripts": {
-    "prepack": "npm run build",
-    "build": "wireit",
-    "build:tsc": "wireit",
-    "dev": "npm run serve --watch",
-    "test": "wireit",
-    "serve": "wireit",
-    "demo:contact": "(cd ../../../agent/adk/contact_lookup && uv run .)",
-    "demo:restaurant": "(cd ../../../agent/adk/restaurant_finder && uv run .)"
-  },
-  "wireit": {
-    "serve": {
-      "command": "vite dev",
-      "dependencies": [
-        "build"
-      ],
-      "service": true
-    },
-    "test": {
-      "command": "node --test --enable-source-maps --test-reporter spec dist/src/0.8/tests/**/*.test.js",
-      "dependencies": [
-        "build"
-      ]
-    },
-    "build": {
-      "dependencies": [
-        "build:tsc"
-      ]
-    },
-    "build:tsc": {
-      "command": "tsc -b --pretty",
-      "env": {
-        "FORCE_COLOR": "1"
-      },
-      "dependencies": [
-        "../../../../renderers/lit:build:tsc"
-      ],
-      "files": [
-        "**/*.ts",
-        "tsconfig.json"
-      ],
-      "output": [
-        "dist/",
-        "!dist/**/*.min.js{,.map}"
-      ],
-      "clean": "if-file-deleted"
-    }
-  },
-  "repository": {
-    "directory": "samples/client/lit/shell_v0_9",
-    "type": "git",
-    "url": "git+https://github.com/google/A2UI.git"
-  },
-  "files": [
-    "dist"
-  ],
-  "keywords": [],
-  "author": "Google",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/google/A2UI/issues"
-  },
-  "homepage": "https://github.com/google/A2UI/tree/main/web#readme",
-  "devDependencies": {
-    "dotenv": "^17.2.3",
-    "typescript": "^5.8.3",
-    "uuid": "^13.0.0",
-    "vite": "^7.1.11",
-    "wireit": "^0.15.0-pre.2"
-  },
-  "dependencies": {
-    "@a2a-js/sdk": "^0.3.4",
-    "@a2ui/lit": "file:../../../../renderers/lit",
-    "@a2ui/markdown-it": "file:../../../../renderers/markdown/markdown-it",
-    "@google/genai": "^1.22.0",
-    "@lit-labs/signals": "^0.1.3",
-    "@lit/context": "^1.1.4",
-    "@types/node": "^24.7.1",
-    "lit": "^3.3.1"
-  }
-}
