--- samples/client/lit/shell_v0_9/tsconfig.json
@@ -1,35 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-
-  "compilerOptions": {
-    "composite": false,
-    "declaration": true,
-    "declarationMap": true,
-    "incremental": true,
-    "forceConsistentCasingInFileNames": true,
-    "inlineSources": false,
-    // "allowJs": true,
-    "preserveWatchOutput": true,
-    "sourceMap": true,
-    "target": "es2022",
-    "module": "es2022",
-    "lib": ["ESNext", "DOM", "DOM.Iterable"],
-    "skipLibCheck": true,
-    "useDefineForClassFields": false,
-    "rootDir": ".",
-    "outDir": "dist",
-    "tsBuildInfoFile": "dist/.tsbuildinfo",
-
-    /* Bundler mode */
-    "moduleResolution": "bundler",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-
-    /* Linting */
-    "strict": false,
-    "noUnusedLocals": false,
-    "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "references": [{ "path": "../../../../renderers/lit" }]
-}
