--- samples/client/lit/shell_v0_9/vite.config.ts
@@ -1,46 +1,0 @@
-/*
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { config } from "dotenv";
-import { UserConfig } from "vite";
-import * as Middleware from "./middleware";
-import { dirname, resolve } from "node:path";
-import { fileURLToPath } from "node:url";
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-export default async () => {
-  config();
-
-  const entry: Record<string, string> = {
-    shell: resolve(__dirname, "index.html"),
-  };
-
-  return {
-    // Do we need this for this sample?
-    plugins: [Middleware.A2AMiddleware.plugin()],
-    build: {
-      rollupOptions: {
-        input: entry,
-      },
-      target: "esnext",
-    },
-    define: {},
-    resolve: {
-      dedupe: ["lit"],
-    },
-  } satisfies UserConfig;
-};
